evaluation/fishfarm/pyproject.toml
[project]
name = "fishfarm"
description = ""
readme = "README.md"
license = {file = "LICENSE"}
authors = [
  {name = "Takuya Akiba"},
  {email = "takiba@sakana.ai"}
]
classifiers = [
    "Development Status :: 2 - Pre-Alpha",
    "Intended Audience :: Science/Research",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.8",
    "Programming Language :: Python :: 3.9",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3 :: Only",
    "Topic :: Scientific/Engineering",
    "Topic :: Scientific/Engineering :: Mathematics",
    "Topic :: Scientific/Engineering :: Artificial Intelligence",
    "Topic :: Software Development",
    "Topic :: Software Development :: Libraries",
    "Topic :: Software Development :: Libraries :: Python Modules",
]
requires-python = ">=3.10"
dependencies = [
    "huggingface_hub",
    "transformers",
    "pydantic",
    "colorlog"
]
dynamic = ["version"]

[project.optional-dependencies]
development = [
    "black",
    "blackdoc",
    "flake8",
    "isort",
    "mypy",
    "pytest",
    "pytest-mock",
    "types-PyYAML",
]

full = [
    "vllm",
    "langchain",
    "langchain-openai",
    "fasttext-wheel",
    "datasets",
    "mysql-connector-python==8.0.32",
    "docker==6.1.2",
    "evalplus @ git+https://github.com/evalplus/evalplus@1895d2f6aa8895044a7cf69defc24bd57695e885",
    "rouge-score"
]

[project.urls]
repository = "https://github.com/SakanaAI/fishfarm"

[tool.setuptools.packages.find]
include = ["fishfarm*"]

[tool.setuptools.dynamic]
version = {attr = "fishfarm.version.__version__"}

[tool.black]
line-length = 99
target-version = ['py310']
exclude = '''
/(
    \.eggs
  | \.git
  | \.hg
  | \.mypy_cache
  | \.venv
  | venv
  | _build
  | buck-out
  | build
  | dist
  | docs
  | data
)/
'''

[tool.isort]
profile = 'black'
src_paths = ['fishfarm', 'tests']
line_length = 99
lines_after_imports = 2

[tool.mypy]
python_version = "3.10"
strict = true
ignore_missing_imports = true
warn_unused_configs = true
disallow_untyped_defs = true
warn_redundant_casts = true
warn_unused_ignores = true
warn_unreachable = true
disallow_any_generics = false
exclude = ".venv|venv|build|docs|tutorial|data"

[tool.pytest]
mock_use_standalone_module = true
